requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
Flask-Login==0.6.2
Werkzeug==2.3.7
python-dotenv==1.0.0
email-validator==2.0.0
gunicorn==20.1.0